test: separate resolver tests into multiple files
