+++ crates/resolver-tests/tests/resolve.rs
@@ -1,278 +1,14 @@
-use std::io::IsTerminal;
-
 use cargo::core::dependency::DepKind;
 use cargo::core::Dependency;
 use cargo::util::GlobalContext;
-use cargo_util::is_ci;
 
 use resolver_tests::{
-    assert_contains, assert_same, dep, dep_kind, dep_loc, dep_req, loc_names, names, pkg, pkg_dep,
-    pkg_dep_with, pkg_id, pkg_loc, registry, registry_strategy, remove_dep, resolve,
-    resolve_and_validated, resolve_with_global_context, PrettyPrintRegistry, SatResolver, ToDep,
-    ToPkgId,
+    helpers::{
+        assert_contains, assert_same, dep, dep_kind, dep_loc, dep_req, loc_names, names, pkg_id,
+        pkg_loc, registry, ToDep, ToPkgId,
+    },
+    pkg, resolve, resolve_with_global_context,
 };
-
-use proptest::prelude::*;
-
-// NOTE: proptest is a form of fuzz testing. It generates random input and makes sure that
-// certain universal truths are upheld. Therefore, it can pass when there is a problem,
-// but if it fails then there really is something wrong. When testing something as
-// complicated as the resolver, the problems can be very subtle and hard to generate.
-// We have had a history of these tests only failing on PRs long after a bug is introduced.
-// If you have one of these test fail please report it on #6258,
-// and if you did not change the resolver then feel free to retry without concern.
-proptest! {
-    #![proptest_config(ProptestConfig {
-        max_shrink_iters:
-            if is_ci() || !std::io::stderr().is_terminal() {
-                // This attempts to make sure that CI will fail fast,
-                0
-            } else {
-                // but that local builds will give a small clear test case.
-                u32::MAX
-            },
-        result_cache: prop::test_runner::basic_result_cache,
-        .. ProptestConfig::default()
-    })]
-
-    /// NOTE: if you think this test has failed spuriously see the note at the top of this macro.
-    #[test]
-    fn prop_passes_validation(
-        PrettyPrintRegistry(input) in registry_strategy(50, 20, 60)
-    )  {
-        let reg = registry(input.clone());
-        let mut sat_resolver = SatResolver::new(&reg);
-
-        // There is only a small chance that a crate will be interesting.
-        // So we try some of the most complicated.
-        for this in input.iter().rev().take(20) {
-            let _ = resolve_and_validated(
-                vec![dep_req(&this.name(), &format!("={}", this.version()))],
-                &reg,
-                &mut sat_resolver,
-            );
-        }
-    }
-
-    /// NOTE: if you think this test has failed spuriously see the note at the top of this macro.
-    #[test]
-    fn prop_minimum_version_errors_the_same(
-            PrettyPrintRegistry(input) in registry_strategy(50, 20, 60)
-    ) {
-        let mut gctx = GlobalContext::default().unwrap();
-        gctx.nightly_features_allowed = true;
-        gctx
-            .configure(
-                1,
-                false,
-                None,
-                false,
-                false,
-                false,
-                &None,
-                &["minimal-versions".to_string()],
-                &[],
-            )
-            .unwrap();
-
-        let reg = registry(input.clone());
-
-        // There is only a small chance that a crate will be interesting.
-        // So we try some of the most complicated.
-        for this in input.iter().rev().take(10) {
-            let deps = vec![dep_req(&this.name(), &format!("={}", this.version()))];
-            let res = resolve(deps.clone(), &reg);
-            let mres = resolve_with_global_context(deps, &reg, &gctx);
-
-            // `minimal-versions` changes what order the candidates are tried but not the existence of a solution.
-            prop_assert_eq!(
-                res.is_ok(),
-                mres.is_ok(),
-                "minimal-versions and regular resolver disagree about whether `{} = \"={}\"` can resolve",
-                this.name(),
-                this.version()
-            )
-        }
-    }
-
-    /// NOTE: if you think this test has failed spuriously see the note at the top of this macro.
-    #[test]
-    fn prop_direct_minimum_version_error_implications(
-            PrettyPrintRegistry(input) in registry_strategy(50, 20, 60)
-    ) {
-        let mut gctx = GlobalContext::default().unwrap();
-        gctx.nightly_features_allowed = true;
-        gctx
-            .configure(
-                1,
-                false,
-                None,
-                false,
-                false,
-                false,
-                &None,
-                &["direct-minimal-versions".to_string()],
-                &[],
-            )
-            .unwrap();
-
-        let reg = registry(input.clone());
-
-        // There is only a small chance that a crate will be interesting.
-        // So we try some of the most complicated.
-        for this in input.iter().rev().take(10) {
-            let deps = vec![dep_req(&this.name(), &format!("={}", this.version()))];
-            let res = resolve(deps.clone(), &reg);
-            let mres = resolve_with_global_context(deps, &reg, &gctx);
-
-            // `direct-minimal-versions` reduces the number of available solutions,
-            //  so we verify that we do not come up with solutions not seen in `maximal-versions`.
-            if res.is_err() {
-                prop_assert!(
-                    mres.is_err(),
-                    "direct-minimal-versions should not have more solutions than the regular, maximal resolver but found one when resolving `{} = \"={}\"`",
-                    this.name(),
-                    this.version()
-                )
-            }
-            if mres.is_ok() {
-                prop_assert!(
-                    res.is_ok(),
-                    "direct-minimal-versions should not have more solutions than the regular, maximal resolver but found one when resolving `{} = \"={}\"`",
-                    this.name(),
-                    this.version()
-                )
-            }
-        }
-    }
-
-    /// NOTE: if you think this test has failed spuriously see the note at the top of this macro.
-    #[test]
-    fn prop_removing_a_dep_cant_break(
-            PrettyPrintRegistry(input) in registry_strategy(50, 20, 60),
-            indexes_to_remove in prop::collection::vec((any::<prop::sample::Index>(), any::<prop::sample::Index>()), ..10)
-    ) {
-        let reg = registry(input.clone());
-        let mut removed_input = input.clone();
-        for (summary_idx, dep_idx) in indexes_to_remove {
-            if !removed_input.is_empty() {
-                let summary_idx = summary_idx.index(removed_input.len());
-                let deps = removed_input[summary_idx].dependencies();
-                if !deps.is_empty() {
-                    let new = remove_dep(&removed_input[summary_idx], dep_idx.index(deps.len()));
-                    removed_input[summary_idx] = new;
-                }
-            }
-        }
-        let removed_reg = registry(removed_input);
-
-        // There is only a small chance that a crate will be interesting.
-        // So we try some of the most complicated.
-        for this in input.iter().rev().take(10) {
-            if resolve(
-                vec![dep_req(&this.name(), &format!("={}", this.version()))],
-                &reg,
-            ).is_ok() {
-                prop_assert!(
-                    resolve(
-                        vec![dep_req(&this.name(), &format!("={}", this.version()))],
-                        &removed_reg,
-                    ).is_ok(),
-                    "full index worked for `{} = \"={}\"` but removing some deps broke it!",
-                    this.name(),
-                    this.version(),
-                )
-            }
-        }
-    }
-
-    /// NOTE: if you think this test has failed spuriously see the note at the top of this macro.
-    #[test]
-    fn prop_limited_independence_of_irrelevant_alternatives(
-        PrettyPrintRegistry(input) in registry_strategy(50, 20, 60),
-        indexes_to_unpublish in prop::collection::vec(any::<prop::sample::Index>(), ..10)
-    )  {
-        let reg = registry(input.clone());
-
-        // There is only a small chance that a crate will be interesting.
-        // So we try some of the most complicated.
-        for this in input.iter().rev().take(10) {
-            let res = resolve(
-                vec![dep_req(&this.name(), &format!("={}", this.version()))],
-                &reg,
-            );
-
-            match res {
-                Ok(r) => {
-                    // If resolution was successful, then unpublishing a version of a crate
-                    // that was not selected should not change that.
-                    let not_selected: Vec<_> = input
-                        .iter()
-                        .cloned()
-                        .filter(|x| !r.contains(&x.package_id()))
-                        .collect();
-
-                    if !not_selected.is_empty() {
-                        let indexes_to_unpublish: Vec<_> = indexes_to_unpublish.iter().map(|x| x.get(&not_selected)).collect();
-
-                        let new_reg = registry(
-                            input
-                                .iter()
-                                .cloned()
-                                .filter(|x| !indexes_to_unpublish.contains(&x))
-                                .collect(),
-                        );
-
-                        let res = resolve(
-                            vec![dep_req(&this.name(), &format!("={}", this.version()))],
-                            &new_reg,
-                        );
-
-                        // Note: that we can not assert that the two `res` are identical
-                        // as the resolver does depend on irrelevant alternatives.
-                        // It uses how constrained a dependency requirement is
-                        // to determine what order to evaluate requirements.
-
-                        prop_assert!(
-                            res.is_ok(),
-                            "unpublishing {:?} stopped `{} = \"={}\"` from working",
-                            indexes_to_unpublish.iter().map(|x| x.package_id()).collect::<Vec<_>>(),
-                            this.name(),
-                            this.version()
-                        )
-                    }
-                }
-
-                Err(_) => {
-                    // If resolution was unsuccessful, then it should stay unsuccessful
-                    // even if any version of a crate is unpublished.
-                    let indexes_to_unpublish: Vec<_> = indexes_to_unpublish.iter().map(|x| x.get(&input)).collect();
-
-                    let new_reg = registry(
-                        input
-                            .iter()
-                            .cloned()
-                            .filter(|x| !indexes_to_unpublish.contains(&x))
-                            .collect(),
-                    );
-
-                    let res = resolve(
-                        vec![dep_req(&this.name(), &format!("={}", this.version()))],
-                        &new_reg,
-                    );
-
-                    prop_assert!(
-                        res.is_err(),
-                        "full index did not work for `{} = \"={}\"` but unpublishing {:?} fixed it!",
-                        this.name(),
-                        this.version(),
-                        indexes_to_unpublish.iter().map(|x| x.package_id()).collect::<Vec<_>>(),
-                    )
-                }
-            }
-        }
-    }
-}
 
 #[test]
 #[should_panic(expected = "assertion failed: !name.is_empty()")]
@@ -1199,214 +935,6 @@ fn large_conflict_cache() {
     }
     let reg = registry(input);
     let _ = resolve(root_deps, &reg);
-}
-
-#[test]
-fn off_by_one_bug() {
-    let input = vec![
-        pkg!(("A-sys", "0.0.1")),
-        pkg!(("A-sys", "0.0.4")),
-        pkg!(("A-sys", "0.0.6")),
-        pkg!(("A-sys", "0.0.7")),
-        pkg!(("NA", "0.0.0") => [dep_req("A-sys", "<= 0.0.5"),]),
-        pkg!(("NA", "0.0.1") => [dep_req("A-sys", ">= 0.0.6, <= 0.0.8"),]),
-        pkg!(("a", "0.0.1")),
-        pkg!(("a", "0.0.2")),
-        pkg!(("aa", "0.0.0") => [dep_req("A-sys", ">= 0.0.4, <= 0.0.6"),dep_req("NA", "<= 0.0.0"),]),
-        pkg!(("f", "0.0.3") => [dep("NA"),dep_req("a", "<= 0.0.2"),dep("aa"),]),
-    ];
-
-    let reg = registry(input);
-    let mut sat_resolver = SatResolver::new(&reg);
-    assert!(resolve_and_validated(vec![dep("f")], &reg, &mut sat_resolver).is_ok());
-}
-
-#[test]
-fn conflict_store_bug() {
-    let input = vec![
-        pkg!(("A", "0.0.3")),
-        pkg!(("A", "0.0.5")),
-        pkg!(("A", "0.0.9") => [dep("bad"),]),
-        pkg!(("A", "0.0.10") => [dep("bad"),]),
-        pkg!(("L-sys", "0.0.1") => [dep("bad"),]),
-        pkg!(("L-sys", "0.0.5")),
-        pkg!(("R", "0.0.4") => [
-            dep_req("L-sys", "= 0.0.5"),
-        ]),
-        pkg!(("R", "0.0.6")),
-        pkg!(("a-sys", "0.0.5")),
-        pkg!(("a-sys", "0.0.11")),
-        pkg!(("c", "0.0.12") => [
-            dep_req("R", ">= 0.0.3, <= 0.0.4"),
-        ]),
-        pkg!(("c", "0.0.13") => [
-            dep_req("a-sys", ">= 0.0.8, <= 0.0.11"),
-        ]),
-        pkg!(("c0", "0.0.6") => [
-            dep_req("L-sys", "<= 0.0.2"),
-        ]),
-        pkg!(("c0", "0.0.10") => [
-            dep_req("A", ">= 0.0.9, <= 0.0.10"),
-            dep_req("a-sys", "= 0.0.5"),
-        ]),
-        pkg!("j" => [
-            dep_req("A", ">= 0.0.3, <= 0.0.5"),
-            dep_req("R", ">=0.0.4, <= 0.0.6"),
-            dep_req("c", ">= 0.0.9"),
-            dep_req("c0", ">= 0.0.6"),
-        ]),
-    ];
-
-    let reg = registry(input);
-    let mut sat_resolver = SatResolver::new(&reg);
-    assert!(resolve_and_validated(vec![dep("j")], &reg, &mut sat_resolver).is_err());
-}
-
-#[test]
-fn conflict_store_more_then_one_match() {
-    let input = vec![
-        pkg!(("A", "0.0.0")),
-        pkg!(("A", "0.0.1")),
-        pkg!(("A-sys", "0.0.0")),
-        pkg!(("A-sys", "0.0.1")),
-        pkg!(("A-sys", "0.0.2")),
-        pkg!(("A-sys", "0.0.3")),
-        pkg!(("A-sys", "0.0.12")),
-        pkg!(("A-sys", "0.0.16")),
-        pkg!(("B-sys", "0.0.0")),
-        pkg!(("B-sys", "0.0.1")),
-        pkg!(("B-sys", "0.0.2") => [dep_req("A-sys", "= 0.0.12"),]),
-        pkg!(("BA-sys", "0.0.0") => [dep_req("A-sys","= 0.0.16"),]),
-        pkg!(("BA-sys", "0.0.1") => [dep("bad"),]),
-        pkg!(("BA-sys", "0.0.2") => [dep("bad"),]),
-        pkg!("nA" => [
-            dep("A"),
-            dep_req("A-sys", "<= 0.0.3"),
-            dep("B-sys"),
-            dep("BA-sys"),
-        ]),
-    ];
-    let reg = registry(input);
-    let mut sat_resolver = SatResolver::new(&reg);
-    assert!(resolve_and_validated(vec![dep("nA")], &reg, &mut sat_resolver).is_err());
-}
-
-#[test]
-fn bad_lockfile_from_8249() {
-    let input = vec![
-        pkg!(("a-sys", "0.2.0")),
-        pkg!(("a-sys", "0.1.0")),
-        pkg!(("b", "0.1.0") => [
-            dep_req("a-sys", "0.1"), // should be optional: true, but not needed for now
-        ]),
-        pkg!(("c", "1.0.0") => [
-            dep_req("b", "=0.1.0"),
-        ]),
-        pkg!("foo" => [
-            dep_req("a-sys", "=0.2.0"),
-            {
-                let mut b = dep_req("b", "=0.1.0");
-                b.set_features(vec!["a-sys"]);
-                b
-            },
-            dep_req("c", "=1.0.0"),
-        ]),
-    ];
-    let reg = registry(input);
-    let mut sat_resolver = SatResolver::new(&reg);
-    assert!(resolve_and_validated(vec![dep("foo")], &reg, &mut sat_resolver).is_err());
-}
-
-#[test]
-fn registry_with_features() {
-    let reg = registry(vec![
-        pkg!("a"),
-        pkg!("b"),
-        pkg_dep_with(
-            "image",
-            vec!["a".to_opt_dep(), "b".to_opt_dep(), "jpg".to_dep()],
-            &[("default", &["a"]), ("b", &["dep:b"])],
-        ),
-        pkg!("jpg"),
-        pkg!("log"),
-        pkg!("man"),
-        pkg_dep_with("rgb", vec!["man".to_opt_dep()], &[("man", &["dep:man"])]),
-        pkg_dep_with(
-            "dep",
-            vec![
-                "image".to_dep_with(&["b"]),
-                "log".to_opt_dep(),
-                "rgb".to_opt_dep(),
-            ],
-            &[
-                ("default", &["log", "image/default"]),
-                ("man", &["rgb?/man"]),
-            ],
-        ),
-    ]);
-
-    for deps in [
-        vec!["dep".to_dep_with(&["default", "man", "log", "rgb"])],
-        vec!["dep".to_dep_with(&["default"])],
-        vec!["dep".to_dep_with(&[])],
-        vec!["dep".to_dep_with(&["man"])],
-        vec!["dep".to_dep_with(&["man", "rgb"])],
-    ] {
-        let mut sat_resolver = SatResolver::new(&reg);
-        assert!(resolve_and_validated(deps, &reg, &mut sat_resolver).is_ok());
-    }
-}
-
-#[test]
-fn missing_feature() {
-    let reg = registry(vec![pkg!("a")]);
-    let mut sat_resolver = SatResolver::new(&reg);
-    assert!(resolve_and_validated(vec!["a".to_dep_with(&["f"])], &reg, &mut sat_resolver).is_err());
-}
-
-#[test]
-fn conflict_feature_and_sys() {
-    let reg = registry(vec![
-        pkg(("a-sys", "1.0.0")),
-        pkg(("a-sys", "2.0.0")),
-        pkg_dep_with(
-            ("a", "1.0.0"),
-            vec![dep_req("a-sys", "1.0.0")],
-            &[("f", &[])],
-        ),
-        pkg_dep_with(
-            ("a", "2.0.0"),
-            vec![dep_req("a-sys", "2.0.0")],
-            &[("g", &[])],
-        ),
-        pkg_dep("dep", vec![dep_req("a", "2.0.0")]),
-    ]);
-
-    let deps = vec![dep_req("a", "*").to_dep_with(&["f"]), dep("dep")];
-    let mut sat_resolver = SatResolver::new(&reg);
-    assert!(resolve_and_validated(deps, &reg, &mut sat_resolver).is_err());
-}
-
-#[test]
-fn conflict_weak_features() {
-    let reg = registry(vec![
-        pkg(("a-sys", "1.0.0")),
-        pkg(("a-sys", "2.0.0")),
-        pkg_dep("a1", vec![dep_req("a-sys", "1.0.0").to_opt_dep()]),
-        pkg_dep("a2", vec![dep_req("a-sys", "2.0.0").to_opt_dep()]),
-        pkg_dep_with(
-            "dep",
-            vec!["a1".to_opt_dep(), "a2".to_opt_dep()],
-            &[("a1", &["a1?/a-sys"]), ("a2", &["a2?/a-sys"])],
-        ),
-    ]);
-
-    let deps = vec![dep("dep").to_dep_with(&["a1", "a2"])];
-
-    // The following asserts should be updated when support for weak dependencies
-    // is added to the dependency resolver.
-    assert!(resolve(deps.clone(), &reg).is_err());
-    assert!(SatResolver::new(&reg).sat_resolve(&deps));
 }
 
 #[test]
